capstone_project/requirements.txt
# Core Framework
streamlit>=1.31.0
python-dotenv>=1.0.1

# Azure OpenAI (GPT-5.1-chat) — direct SDK
openai>=1.10.0

# Machine Learning
scikit-learn>=1.4.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.3
openpyxl>=3.1.2

# Visualization
plotly>=5.18.0
matplotlib>=3.8.2
seaborn>=0.13.1

# Statistics (for trendlines in plotly)
statsmodels>=0.14.0
scipy>=1.12.0

# Utilities
pydantic>=2.5.3
python-dateutil>=2.8.2
cachetools>=5.3.2
